+++ src/duration.rs
@@ -13,16 +13,96 @@ mod macros;
 
 /// Represents a duration of time.
 ///
-/// The generic `T` can either be `u32` or `u64`, typenums represent the Period in seconds
+/// The generic `T` can either be `u32` or `u64`, though progress is made make it `PrimInt.`
 /// Duration in seconds is calculated as `Numer / Denom * ticks`
 #[derive(Clone, Copy, Debug)]
-pub struct Duration<T: PrimInt, Numer, Denom: NonZero> {
+pub struct Duration<T, Numer, Denom>
+where
+    T: PrimInt,
+    Denom: NonZero,
+{
     pub(crate) ticks: T,
     pub(crate) _period: Period<Numer, Denom>,
 }
 impl_duration_for_integer!(u32);
 impl_duration_for_integer!(u64);
 
+impl<T, Numer, Denom> Duration<T, Numer, Denom>
+where
+    T: PrimInt,
+    Numer: Unsigned,
+    Denom: Unsigned + NonZero,
+{
+    /// Create a `Duration` from a ticks value.
+    ///
+    /// ```
+    /// # use typus_fugit::*;
+    /// let _d = Duration::<u32, typenum::U1, typenum::U1000>::from_ticks(1);
+    /// ```
+    #[inline]
+    pub const fn from_ticks(ticks: T) -> Self {
+        Self {
+            ticks,
+            _period: Period {
+                _numer: PhantomData,
+                _denom: PhantomData,
+            },
+        }
+    }
+
+    /// Extract the ticks from a `Duration`.
+    ///
+    /// ```
+    /// # use typus_fugit::*;
+    /// let d = Duration::<u32, typenum::U1, typenum::U1000>::from_ticks(234);
+    ///
+    /// assert_eq!(d.ticks(), 234);
+    /// ```
+    #[inline]
+    pub const fn ticks(&self) -> T {
+        self.ticks
+    }
+}
+
+// Duration - Duration = Duration (only same base until const_generics_defaults is
+// stabilized)
+// UPDATE v0.4.0: With `typenum`, this should now be implementable
+impl<ST, OT, Numer, Denom> ops::Sub<Duration<OT, Numer, Denom>> for Duration<ST, Numer, Denom>
+where
+    ST: PrimInt + From<OT>,
+    OT: PrimInt,
+    Numer: Unsigned,
+    Denom: Unsigned + NonZero,
+{
+    type Output = Duration<ST, Numer, Denom>;
+
+    #[inline]
+    fn sub(self, other: Duration<OT, Numer, Denom>) -> Self::Output {
+        if let Some(v) = self.ticks.checked_sub(&other.ticks.into()) {
+            Self {
+                ticks: v,
+                _period: Period::default(),
+            }
+        } else {
+            panic!("Sub failed!");
+        }
+    }
+}
+
+// Duration -= Duration
+impl<ST, OT, Numer, Denom> ops::SubAssign<Duration<OT, Numer, Denom>> for Duration<ST, Numer, Denom>
+where
+    ST: PrimInt,
+    OT: PrimInt,
+    Self: ops::Sub<Duration<OT, Numer, Denom>, Output = Self>,
+    Numer: Unsigned,
+    Denom: Unsigned + NonZero,
+{
+    #[inline]
+    fn sub_assign(&mut self, other: Duration<OT, Numer, Denom>) {
+        *self = *self - other;
+    }
+}
 //
 // Operations between u32 and u64 Durations
 //
@@ -53,57 +133,25 @@ where
     }
 }
 
-// Duration - Duration = Duration (to make shorthands work, until const_generics_defaults is
-// stabilized)
-// UPDATE v0.4.0: With `typenum`, this should now be implementable
-impl<Numer, Denom> ops::Sub<Duration<u32, Numer, Denom>> for Duration<u64, Numer, Denom>
-where
-    Numer: Unsigned,
-    Denom: Unsigned + NonZero,
-{
-    type Output = Duration<u64, Numer, Denom>;
-
-    #[inline]
-    fn sub(self, other: Duration<u32, Numer, Denom>) -> Self::Output {
-        if let Some(v) = self.checked_sub(Duration::<u64, Numer, Denom>::from_ticks(
-            other.ticks() as u64
-        )) {
-            v
-        } else {
-            panic!("Sub failed!");
-        }
-    }
-}
-
-// Duration -= Duration (to make shorthands work, until const_generics_defaults is stabilized)
-// UPDATE v0.4.0: With `typenum`, this should now be implementable
-impl<Numer, Denom> ops::SubAssign<Duration<u32, Numer, Denom>> for Duration<u64, Numer, Denom>
-where
-    Numer: Unsigned,
-    Denom: Unsigned + NonZero,
-{
-    #[inline]
-    fn sub_assign(&mut self, other: Duration<u32, Numer, Denom>) {
-        *self = *self - other;
-    }
-}
-
 // Duration + Duration = Duration (to make shorthands work, until const_generics_defaults is
 // stabilized)
 // UPDATE v0.4.0: With `typenum`, this should now be implementable
-impl<Numer, Denom> ops::Add<Duration<u32, Numer, Denom>> for Duration<u64, Numer, Denom>
+impl<ST, OT, Numer, Denom> ops::Add<Duration<OT, Numer, Denom>> for Duration<ST, Numer, Denom>
 where
+    ST: PrimInt + From<OT>,
+    OT: PrimInt,
     Numer: Unsigned,
     Denom: Unsigned + NonZero,
 {
-    type Output = Duration<u64, Numer, Denom>;
+    type Output = Duration<ST, Numer, Denom>;
 
     #[inline]
-    fn add(self, other: Duration<u32, Numer, Denom>) -> Self::Output {
-        if let Some(v) = self.checked_add(Duration::<u64, Numer, Denom>::from_ticks(
-            other.ticks() as u64
-        )) {
-            v
+    fn add(self, other: Duration<OT, Numer, Denom>) -> Self::Output {
+        if let Some(v) = self.ticks.checked_add(&other.ticks().into()) {
+            Self {
+                ticks: v,
+                _period: Period::default(),
+            }
         } else {
             panic!("Add failed!");
         }
@@ -112,13 +160,15 @@ where
 
 // Duration += Duration (to make shorthands work, until const_generics_defaults is stabilized)
 // UPDATE v0.4.0: With `typenum`, this should now be implementable
-impl<Numer, Denom> ops::AddAssign<Duration<u32, Numer, Denom>> for Duration<u64, Numer, Denom>
+impl<ST, OT, Numer, Denom> ops::AddAssign<Duration<OT, Numer, Denom>> for Duration<ST, Numer, Denom>
 where
+    ST: PrimInt + From<OT>,
+    OT: PrimInt,
     Numer: Unsigned,
     Denom: Unsigned + NonZero,
 {
     #[inline]
-    fn add_assign(&mut self, other: Duration<u32, Numer, Denom>) {
+    fn add_assign(&mut self, other: Duration<OT, Numer, Denom>) {
         *self = *self + other;
     }
 }