+++ src/rate.rs
@@ -13,9 +13,9 @@ mod macros;
 /// The generic `T` can either be `u32` or `u64`, and the const generics represent the ratio of the
 /// raw contained within the rate: `rate in Hz = Numer / Denom * raw`
 #[derive(Clone, Copy, Debug)]
-pub struct Rate<T: PrimInt, Numer, Denom>
+pub struct Rate<T, Numer, Denom>
 where
-    Numer: Unsigned,
+    T: PrimInt,
     Denom: Unsigned + NonZero,
 {
     pub(crate) raw: T,
@@ -26,6 +26,106 @@ where
 impl_rate_for_integer!(u32);
 impl_rate_for_integer!(u64);
 
+impl<T, Numer, Denom> Rate<T, Numer, Denom>
+where
+    T: PrimInt,
+    Numer: Unsigned,
+    Denom: Unsigned + NonZero,
+{
+    /// Create a `Rate` from a raw value.
+    ///
+    /// ```
+    /// # use typus_fugit::*;
+    /// let _d = Rate::<_, typenum::U1, typenum::U1000>::from_raw(1);
+    /// ```
+    #[inline]
+    pub const fn from_raw(raw: T) -> Self {
+        Rate {
+            raw,
+            _numer: PhantomData,
+            _denom: PhantomData,
+        }
+    }
+
+    /// Extract the raw value from a `Rate`.
+    ///
+    /// ```
+    /// # use typus_fugit::*;
+    /// let d = Rate::<_, typenum::U1, typenum::U1000>::from_raw(234);
+    ///
+    /// assert_eq!(d.raw(), 234);
+    /// ```
+    #[inline]
+    pub const fn raw(&self) -> T {
+        self.raw
+    }
+}
+// Rate - Rate = Rate (only same base until const_generics_defaults is
+// stabilized)
+// UPDATE v0.4.0: With `typenum`, this should now be implementable
+impl<ST, OT, Numer, Denom> ops::Sub<Rate<OT, Numer, Denom>> for Rate<ST, Numer, Denom>
+where
+    ST: PrimInt + From<OT>,
+    OT: PrimInt,
+    Numer: Unsigned,
+    Denom: Unsigned + NonZero,
+{
+    type Output = Rate<ST, Numer, Denom>;
+
+    #[inline]
+    fn sub(self, other: Rate<OT, Numer, Denom>) -> Self::Output {
+        if let Some(v) = self.raw.checked_sub(&other.raw.into()) {
+            Self {
+                raw: v,
+                _numer: PhantomData,
+                _denom: PhantomData,
+            }
+        } else {
+            panic!("Sub failed!");
+        }
+    }
+}
+
+// Rate + Rate = Rate (only same base until const_generics_defaults is
+// stabilized)
+// UPDATE v0.4.0: With `typenum`, this should now be implementable
+impl<ST, OT, Numer, Denom> ops::Add<Rate<OT, Numer, Denom>> for Rate<ST, Numer, Denom>
+where
+    ST: PrimInt + From<OT>,
+    OT: PrimInt,
+    Numer: Unsigned,
+    Denom: Unsigned + NonZero,
+{
+    type Output = Rate<ST, Numer, Denom>;
+
+    #[inline]
+    fn add(self, other: Rate<OT, Numer, Denom>) -> Self::Output {
+        if let Some(v) = self.raw.checked_add(&other.raw.into()) {
+            Self {
+                raw: v,
+                _numer: PhantomData,
+                _denom: PhantomData,
+            }
+        } else {
+            panic!("Add failed!");
+        }
+    }
+}
+
+// Rate += Rate
+impl<ST, OT, Numer, Denom> ops::AddAssign<Rate<OT, Numer, Denom>> for Rate<ST, Numer, Denom>
+where
+    ST: PrimInt + From<OT>,
+    OT: PrimInt,
+    Self: ops::Add<Rate<OT, Numer, Denom>, Output = Self>,
+    Numer: Unsigned,
+    Denom: Unsigned + NonZero,
+{
+    #[inline]
+    fn add_assign(&mut self, other: Rate<OT, Numer, Denom>) {
+        *self = *self + other;
+    }
+}
 //
 // Operations between u32 and u64 Rate
 //
@@ -56,26 +156,6 @@ where
     }
 }
 
-// Rate - Rate = Rate (to make shorthands work, until const_generics_defaults is
-// stabilized)
-// UPDATE v0.4.0: With `typenum`, this should now be implementable
-impl<Numer, Denom> ops::Sub<Rate<u32, Numer, Denom>> for Rate<u64, Numer, Denom>
-where
-    Numer: Unsigned,
-    Denom: Unsigned + NonZero,
-{
-    type Output = Rate<u64, Numer, Denom>;
-
-    #[inline]
-    fn sub(self, other: Rate<u32, Numer, Denom>) -> Self::Output {
-        if let Some(v) = self.checked_sub(Rate::<u64, Numer, Denom>::from_raw(other.raw() as u64)) {
-            v
-        } else {
-            panic!("Sub failed!");
-        }
-    }
-}
-
 // Rate -= Rate (to make shorthands work, until const_generics_defaults is stabilized)
 // UPDATE v0.4.0: With `typenum`, this should now be implementable
 impl<Numer, Denom> ops::SubAssign<Rate<u32, Numer, Denom>> for Rate<u64, Numer, Denom>
@@ -86,39 +166,6 @@ where
     #[inline]
     fn sub_assign(&mut self, other: Rate<u32, Numer, Denom>) {
         *self = *self - other;
-    }
-}
-
-// Rate + Rate = Rate (to make shorthands work, until const_generics_defaults is
-// stabilized)
-// UPDATE v0.4.0: With `typenum`, this should now be implementable
-impl<Numer, Denom> ops::Add<Rate<u32, Numer, Denom>> for Rate<u64, Numer, Denom>
-where
-    Numer: Unsigned,
-    Denom: Unsigned + NonZero,
-{
-    type Output = Rate<u64, Numer, Denom>;
-
-    #[inline]
-    fn add(self, other: Rate<u32, Numer, Denom>) -> Self::Output {
-        if let Some(v) = self.checked_add(Rate::<u64, Numer, Denom>::from_raw(other.raw() as u64)) {
-            v
-        } else {
-            panic!("Add failed!");
-        }
-    }
-}
-
-// Rate += Rate (to make shorthands work, until const_generics_defaults is stabilized)
-// UPDATE v0.4.0: With `typenum`, this should now be implementable
-impl<Numer, Denom> ops::AddAssign<Rate<u32, Numer, Denom>> for Rate<u64, Numer, Denom>
-where
-    Numer: Unsigned,
-    Denom: Unsigned + NonZero,
-{
-    #[inline]
-    fn add_assign(&mut self, other: Rate<u32, Numer, Denom>) {
-        *self = *self + other;
     }
 }